feat: add buildGoalRanks helper
