+++ backend/services/lesson/index.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import { runPython } from '../../shared/utils';
+import { runPython, buildGoalRanks } from '../../shared/utils';
 import {
   loadGoals,
   loadReviewState,
@@ -21,18 +21,7 @@ export function createLessonService() {
     const goals = await loadGoals();
     const defaults = await loadDefaultCocaWords();
     const review = await loadReviewState();
-    const customGoals = goals.filter((g) => !g.is_default);
-    const activeGoals =
-      customGoals.length > 0
-        ? customGoals
-        : defaults.map((w) => ({ word: w }));
-    const goalRanks: Record<string, number> = {};
-    defaults.forEach((w, i) => {
-      goalRanks[w] = i + 1;
-    });
-    activeGoals.forEach((g, i) => {
-      goalRanks[g.word] = i + 1;
-    });
+    const { activeGoals, goalRanks } = buildGoalRanks(goals, defaults, review);
     const code = `
 import json, sys
 from datetime import datetime
@@ -99,18 +88,7 @@ print(json.dumps({'lesson': lesson, 'words': list(dict.fromkeys(new_words+review
     const goals = await loadGoals();
     const defaults = await loadDefaultCocaWords();
     const state = await loadReviewState();
-    const customGoals = goals.filter((g) => !g.is_default);
-    const activeGoals =
-      customGoals.length > 0
-        ? customGoals
-        : defaults.map((w) => ({ word: w }));
-    const goalRanks: Record<string, number> = {};
-    defaults.forEach((w, i) => {
-      goalRanks[w] = i + 1;
-    });
-    activeGoals.forEach((g, i) => {
-      goalRanks[g.word] = i + 1;
-    });
+    const { goalRanks } = buildGoalRanks(goals, defaults, state);
     const code = `
 import json, sys
 from datetime import datetime
