+++ backend/shared/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './redis';
 export * from './python';
+export * from './goals';
